Product Work 4
MAINMVC DONE.
DATAFILTERWINDOW DONE.
Made additions to the necessary methods for various other classes.
Fixed a problem with the empty DataSet class in order to remove a very
bothersome error I was getting.
Added more JavaDoc comments, and added TODO comments to other classes
that need to be filled with JavaDoc comments.

As far as MVCs are concerned, this project's missing the final bits of
work on the ToolMVC classes, and even those are halfway done. This will
be the state in which I'll present the work done for this week to the
professor.