Product Work 5
Program runs! Does a whole lot of nothing at the moment, though. Next
step is to debug it, and that'll be done this weekend, mostly.

Interpolation and Outlier Searching is not implemented, nor are DataSets
implemented. (Just lots of empty methods!) Saving, Loading, and Editing
Data / Templates / Data Filters are (mostly, due to the lack of DataSet
implementation) working, though.