Product Work 8.6
Changed DataSet and DataColumn AGAIN. This time, DataSet contains an ArrayList<DataColumn>, and DataColumn contains an ArrayList<E>.
E is the type that will be inserted into the array. However, due to java limitations without workarounds that I can find in a single evening that actually work without errors, I'm going to require all initializations of it to include a String that contains the type as a name.
Currently, only the words "double", "Double", "string", and "String" are allowed.