Error Handling, Exceptions, and Controllers
Adding more JOptionPane error panes via the ExceptionHandler class.
Trying to keep all exception-handling try-catches in the Controllers of the MVC.
Need to provide adequate interfaces between file reading in MainModel and what appears elsewhere.
Might end up having a visual glitch in this, too. If so, it's the same as the template thing.