Graphing Modifications 1
Added changes to DataSet and DataReference. Now, DataSet mimics Template in that there will only be one of each at a time in the program.
This is because of the Import Data option available to the user. Now, DataSet contains Event-firing that will throw information to the DataSetView jLists when needed.
A few other methods were changed to adapt to this style of graph data management, including the createPair method, which was split up into the createXYGraphPair and createBarGraphPair methods.
However, there are errors in two tests: ReferenceTests and InterpolationTest. These must be fixed!